+++ web/packages/new/photos/services/ml/index.ts
@@ -14,7 +14,7 @@ import { FileType } from "@/media/file-type";
 import { ensure } from "@/utils/ensure";
 import { throttled } from "@/utils/promise";
 import { proxy, transfer } from "comlink";
-import { getRemoteFlag, updateRemoteFlag } from "../settings";
+import { getRemoteFlag, updateRemoteFlag } from "../remote-store";
 import { setSearchPeople } from "../search";
 import type { UploadItem } from "../upload/types";
 import {
